Change weight representation in BOWL to double precision.

+++ psl-core/src/main/java/org/linqs/psl/application/learning/weight/search/bayesian/GaussianProcessKernel.java
@@ -48,19 +48,19 @@ public abstract class GaussianProcessKernel {
 
     /**
       * Compute the kernel, but use buffers provided by the caller.
-      * The use of the buffers must be theadsafe and there are no requirements or guarentees on the
+      * The use of the buffers must be theadsafe and there are no requirements or guarantees on the
       * contents of the buffers before/after invocation (except that they are sized
       * the same as the points).
       * The matrices will be modified.
       */
-    public float kernel(float[] point1, float[] point2, float[] buffer1, float[] buffer2, FloatMatrix matrixShell1, FloatMatrix matrixShell2) {
+    public float kernel(double[] point1, double[] point2, float[] buffer1, float[] buffer2, FloatMatrix matrixShell1, FloatMatrix matrixShell2) {
         assert(point1.length == point2.length);
         assert(buffer1.length == buffer2.length);
         assert(point1.length == buffer1.length);
 
         for (int i = 0; i < point1.length; i++) {
-            buffer1[i] = point1[i];
-            buffer2[i] = point2[i];
+            buffer1[i] = (float)point1[i];
+            buffer2[i] = (float)point2[i];
         }
 
         matrixShell1.assume(buffer1, buffer1.length, 1);
@@ -72,7 +72,7 @@ public abstract class GaussianProcessKernel {
     /**
       * Compute the kernels, but allocate new buffer for the computation.
       */
-    public float kernel(float[] point1, float[] point2) {
+    public float kernel(double[] point1, double[] point2) {
         return kernel(point1, point2, new float[point1.length], new float[point2.length], new FloatMatrix(), new FloatMatrix());
     }
 }
