+++ psl-core/src/test/java/org/linqs/psl/application/learning/weight/search/bayesian/SquaredExpKernelTest.java
@@ -12,8 +12,8 @@ public class SquaredExpKernelTest {
     public void testSquaredExpKernel() {
         Options.WLA_GPP_KERNEL_REL_DEP.set(1.0f);
 
-        float[] x = {1.0f, 2.0f, 3.0f, 4.0f};
-        float[] y = {3.0f, 4.0f, 5.0f, 6.0f};
+        double[] x = {1.0f, 2.0f, 3.0f, 4.0f};
+        double[] y = {3.0f, 4.0f, 5.0f, 6.0f};
 
         GaussianProcessKernel kernel = null;
